CSS Fixed
I fixed the CSS issue with the file. To summarize, I added an open and close <head> and <body >tag around the entire code to structure the html, including the script. Then for the CSS, the issue was with the undefined CSS variables so I found the global styles from Teleport HQ and assigned the variables to the CSS code. Please take the opportunity to run the code on your end and start a pull request if you make any changes.